requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
chromadb==0.5.23
sentence-transformers==3.3.1
python-docx==1.1.2
python-dotenv==1.0.1
google-genai==1.0.0

# Memory optimization dependencies
psutil==6.1.1

# Load testing (optional)
locust==2.32.4
requests==2.32.3
